refactor: move activity to ui folder
